@@ -10,13 +10,14 @@ homepage = "https://github.com/spgi-ci/spgci-python"
 documentation = "https://developer.spglobal.com/commodityinsights/tools/notebooks"
 
 [tool.poetry.dependencies]
-python = "^3.7.0"
+python = "^3.8.0"
 pandas = ">= 1.1.5"
 requests = "^2.27.1"
 tqdm = "^4.64.1"
 typing-extensions = "^4.4.0"
 tenacity = "^8.2.2"
 urllib3 = "1.26.15"
+packaging = "^24.1"
 
 
 
